requirements.txt
flask
pandas
openai>=1.2.3
python-docx
nltk
openpyxl
werkzeug
gunicorn
requests
anthropic>=0.25.0
google-generativeai>=0.3.0
